WIP API public of the render
