+++ package.json
@@ -17,6 +17,7 @@
   ],
   "scripts": {
     "build": "npm-run-all --silent rollup:umd -p minify:umd -s size",
+    "dev": "npm-run-all --silent rollup:umd",
     "clean": "rimraf build && mkdirp build",
     "rollup:umd": "rollup -c",
     "minify:umd": "uglifyjs build/fiber.js -cm -o build/fiber.js",
